pyproject.toml
[tool.poetry]
name = "stl2svg"
version = "0.1.0"
description = ""
authors = ["Fab <fabian@fabianhaupt.de>"]
readme = "README.md"
packages = [{include = "stl2svg", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
svgwrite = "^1.4.3"
numpy = "^2.2.1"
numpy-stl = "^3.2.0"
scipy = "^1.15.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
stl2svg = "stl2svg.core:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q --cov=src/stl2svg" 

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
